switched to using double rather BigDecimal due to interop web service issues.

+++ src/main/java/com/leon/ws/rfq/option/model/OptionPricingController.java
@@ -3,40 +3,39 @@ package com.leon.ws.rfq.option.model;
 import javax.jws.WebService;
 import javax.jws.WebMethod;
 import javax.jws.WebParam;
-import java.math.BigDecimal;
 
 
 @WebService(name="OptionPricingController")
 public interface OptionPricingController
 {
 	@WebMethod
-	void parameterize(@WebParam(name="strike") BigDecimal strike,
-			@WebParam(name="volatility") BigDecimal volatility,
-			@WebParam(name="underlyingPrice") BigDecimal underlyingPrice,
-			@WebParam(name="daysToExpiry") BigDecimal daysToExpiry,
-			@WebParam(name="interestRate") BigDecimal interestRate,
+	void parameterize(@WebParam(name="strike") double strike,
+			@WebParam(name="volatility") double volatility,
+			@WebParam(name="underlyingPrice") double underlyingPrice,
+			@WebParam(name="daysToExpiry") double daysToExpiry,
+			@WebParam(name="interestRate") double interestRate,
 			@WebParam(name="isCall") boolean isCall,
 			@WebParam(name="isEuropean") boolean isEuropean,
-			@WebParam(name="dayCountConvention") BigDecimal dayCountConvention);
+			@WebParam(name="dayCountConvention") double dayCountConvention);
 	
 	
 	@WebMethod
-	OptionPriceResult calculate(@WebParam(name="strike") BigDecimal strike,
-			@WebParam(name="volatility") BigDecimal volatility,
-			@WebParam(name="underlyingPrice") BigDecimal underlyingPrice,
-			@WebParam(name="daysToExpiry") BigDecimal daysToExpiry,
-			@WebParam(name="interestRate") BigDecimal interestRate,
+	OptionPriceResult calculate(@WebParam(name="strike") double strike,
+			@WebParam(name="volatility") double volatility,
+			@WebParam(name="underlyingPrice") double underlyingPrice,
+			@WebParam(name="daysToExpiry") double daysToExpiry,
+			@WebParam(name="interestRate") double interestRate,
 			@WebParam(name="isCall") boolean isCall,
 			@WebParam(name="isEuropean") boolean isEuropean,
-			@WebParam(name="dayCountConvention") BigDecimal dayCountConvention);
+			@WebParam(name="dayCountConvention") double dayCountConvention);
 	
 	@WebMethod
 	String getModelDetails();	
 	
 	@WebMethod
 	OptionPriceResultSet calculateRange(@WebParam(name="rangeKey") String rangeKey,
-			@WebParam(name="startValue") BigDecimal startValue,
-			@WebParam(name="endValue") BigDecimal endValue,
-			@WebParam(name="increment") BigDecimal increment); 
+			@WebParam(name="startValue") double startValue,
+			@WebParam(name="endValue") double endValue,
+			@WebParam(name="increment") double increment); 
 	
 }
